CustomerManagement/CustomerApp/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-cors-headers==4.3.1
django-mysql==4.13.0
djangorestframework==3.15.1
mysqlclient==2.2.4
pika==1.3.2
Pygments==2.17.2
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
